+++ packages/plugins/externals/kanban/vite.config.ts
@@ -8,6 +8,7 @@ export default defineConfig({
 		vettaPluginFederation({
 			name: "kanban",
 			entry: "./src/index.tsx",
+			hostUi: true,
 			hostThemeUi: true,
 		}),
 	],